add support to reset clients without restarting the server
