+++ server/utils/clientFactory.js
@@ -14,6 +14,7 @@ const assert = require('assert');
 const Config = require('bcfg');
 
 const logger = require('../logger');
+const { loadClientConfigs, createConfigsMap } = require('./configs');
 
 const logClientInfo = (id, type, { ssl, host, port, network }) =>
   logger.info(
@@ -136,4 +137,33 @@ function clientFactory(config) {
   return { nodeClient, walletClient, multisigWalletClient };
 }
 
-module.exports = clientFactory;
+/*
+ * Build a map of all clients.
+ * @param {bcfg.Config} config - the main app config object
+ * @returns {Object} - a map of the configs and the clients
+ */
+function buildClients(config) {
+  const { loadClientConfigs, createConfigsMap } = require('./configs');
+  // loadConfigs uses the bpanelConfig to find the clients and build
+  // each of their configs.
+  const clientConfigs = loadClientConfigs(config);
+  const configsMap = createConfigsMap(clientConfigs);
+  if (!clientConfigs.length)
+    logger.warn(
+      'Could not find any client config files. You can use the Connection Manager \
+      settings plugin or read more about configuring connections in the documentation: \
+      https://bpanel.org/docs/configuration.html'
+    );
+  const clients = clientConfigs.reduce((clientsMap, cfg) => {
+    const id = cfg.str('id');
+    assert(id, 'client config must have id');
+    clientsMap.set(id, { ...clientFactory(cfg), config: cfg });
+    return clientsMap;
+  }, new Map());
+  return { configsMap, clients };
+}
+
+module.exports = {
+  clientFactory,
+  buildClients
+};
