+++ server/utils/index.js
@@ -6,7 +6,8 @@
 
 exports.npmExists = require('./npm-exists');
 exports.configHelpers = require('./configs');
-exports.clientFactory = require('./clientFactory');
+exports.clientFactory = require('./clientFactory').clientFactory;
+exports.buildClients = require('./clientFactory').buildClients;
 exports.attach = require('./attach');
 exports.apiFilters = require('./apiFilters');
 exports.pluginUtils = require('./plugins');
